feat(app): 统一命令框架 + 工程化加固（崩溃后门/编码/主机单测/可复现构建）
- cmd_shell 统一命令注册表（UART/TCP/CAN 传输位），tcp_svc 重构为 TCP 适配器
- CRASH_INJECT_ENABLE 跟随 APP_DEBUG_MODE：发布构建自动禁用崩溃注入
- pinout.h/err_mgr.c/shell.c 编码归一为 UTF-8 并恢复损坏注释
- CMakeLists 增加 CMAKE_CROSSCOMPILING 分支：纯主机构建可跑协议 ctest；
  gen_cmake.py 模板同步；cmake/APP.ld + 工具链文件入库保证可复现
- logger 输出路由（LOG_SetSink）、shell ESC/历史/补全增强
- 验证：Keil 0 Error/0 Warning；GCC 构建通过（text≈197KB）；ctest 全绿

+++ APP/APP/BSP/LCD/lcd.c
@@ -1073,9 +1073,9 @@ void lcd_show_char(uint16_t x, uint16_t y, char chr, uint8_t size, uint8_t mode,
 
     chr = (char)(chr - ' ');
     switch (size) {
-        case 12: pfont = (uint8_t *)asc2_1206[chr]; w = 6;  h = 12; bpc = 2; break;
-        case 16: pfont = (uint8_t *)asc2_1608[chr]; w = 8;  h = 16; bpc = 2; break;
-        case 24: pfont = (uint8_t *)asc2_2412[chr]; w = 12; h = 24; bpc = 3; break;
+        case 12: pfont = (uint8_t *)asc2_1206[(uint8_t)chr]; w = 6;  h = 12; bpc = 2; break;
+        case 16: pfont = (uint8_t *)asc2_1608[(uint8_t)chr]; w = 8;  h = 16; bpc = 2; break;
+        case 24: pfont = (uint8_t *)asc2_2412[(uint8_t)chr]; w = 12; h = 24; bpc = 3; break;
         default: return;
     }
     if ((uint32_t)x + w > lcddev.width ||
